requirements/devel.txt
djangorestframework==3.5.1
freezegun==0.3.7
pytest-django==3.0.0
pytest-mock==1.2
pytest==3.0.3
python-dateutil==2.5.3
